refactor: extract LegalHoldStatus to its own file
